// web/src/pages/proposals/students/students.type.ts
import type { ProposalStatus } from "@/lib/utils";

export type ProposalDetail = {
	id: number;
	title: string;
	slug: string;
	description: string;
	file: string;
	status: ProposalStatus;
	type: string;
	projectType: string;
	submittedAt: string;
	supervisor: {
		id: number;
		name: string;
		email: string;
	};
	submittedBy: {
		id: number;
		name: string;
		email: string;
	};
	members: Array<{
		id: number;
		name: string;
		email: string;
	}>;
};